[CNP-111] File names refactoring
