+++ src/app.module.ts
@@ -6,7 +6,6 @@ import { ConfigModule, ConfigService } from '@nestjs/config';
 import { SequelizeModule } from '@nestjs/sequelize';
 import { RolesModule } from './roles/roles.module';
 import { EmployeesModule } from './employees/employees.module';
-import { LegalRepsModule } from './legalReps/legalReps.module';
 import { CustomersModule } from './customers/customers.module';
 import { ProvidersModule } from './providers/providers.module';
 import { QuotationReqsModule } from './quotationReqs/quotationReqs.module';
@@ -16,7 +15,6 @@ import { QuotationServicesModule } from './quotation-services/quotation-services
 import { SessionModule } from './session/session.module';
 import { InvoicesModule } from './invoices/invoices.module';
 import { QuoteReqDetailsModule } from './quoteReqDetails/quoteReqDetails.module';
-import { PackageDetail } from './packageDetails/models/packageDetails.model';
 import { PackageDetailModule } from './packageDetails/packageDetails.module';
 import { FreightModule } from './freight/freight.module';
 import { AirFreightModule } from './airFreight/airFreight.module';
@@ -28,6 +26,7 @@ import { ContractsModule } from './contracts/contracts.module';
 import { ShipmentModule } from './shipment/shipment.module';
 import { DocumentModule } from './document/document.module';
 import { ShipmentTrackingModule } from './shipment-tracking/shipment-tracking.module';
+import { LegalRepsModule } from './legal-representative/legal-rep.module';
 
 @Module({
   imports: [
