+++ src/customers/strategies/create-customer/create-customer-strategy.interface.ts
@@ -1,4 +1,4 @@
-import { CreateCustomerDto } from '@/customers/dtos/CreateCustomerDto';
+import { CreateCustomerDto } from '@/customers/dtos/create-customer.dto';
 import { Customer } from '@/customers/models/customer.model';
 
 export interface ICreateCustomerStrategy {
