+++ src/customers/strategies/find-customer/find-all.strategy.ts
@@ -1,6 +1,5 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { Customer } from '@/customers/models/customer.model';
-import { LegalRep } from '@/legalReps/models/legalReps.model';
 
 @Injectable()
 export class FindAllCustomerStrategy {
