+++ src/legal-representative/dtos/query-legal-rep.dto.ts
