+++ src/legal-representative/legal-rep.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { SequelizeModule } from '@nestjs/sequelize';
-import { LegalRepsService } from './legalReps.service';
-import { LegalRep } from './models/legalReps.model';
-import { LegalRepsController } from './legalReps.controller';
+import { LegalRepsService } from './legal-rep.service';
+import { LegalRep } from './models/legal-rep.model';
+import { LegalRepsController } from './legal-rep.controller';
 import { CreateLegalRepsStrategy } from './strategies/create-legal-rep/create-legal-rep.strategy';
 import { UpdateLegalRepsStrategy } from './strategies/update-legal-rep/update-legal-rep.strategy';
 import { FindAllLegalRepsStrategy } from './strategies/find-legal-rep/find-all.strategy';
