+++ src/legal-representative/strategies/find-legal-rep/find-all.strategy.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { IFindLegalRepsStrategy } from './find-legal-rep-strategy.interface';
-import { LegalRep } from '@/legalReps/models/legalReps.model';
+import { LegalRep } from '@/legal-representative/models/legal-rep.model';
 
 @Injectable()
 export class FindAllLegalRepsStrategy implements IFindLegalRepsStrategy {
