+++ src/legal-representative/strategies/find-legal-rep/find-legal-rep-strategy.interface.ts
@@ -1,4 +1,4 @@
-import { LegalRep } from '@/legalReps/models/legalReps.model';
+import { LegalRep } from '@/legal-representative/models/legal-rep.model';
 
 export interface IFindLegalRepsStrategy {
   find(legalRepInfo: string): Promise<LegalRep[] | null>;
