+++ src/legal-representative/strategies/update-legal-rep/update-legal-rep-strategy.interface.ts
@@ -1,0 +1,9 @@
+import { CreateLegalRepDto } from '@/legal-representative/dtos/create-legal-rep.dto';
+import { LegalRep } from '@/legal-representative/models/legal-rep.model';
+
+export interface IUpdateLegalRepsStrategy {
+  update(
+    legalRepId: string,
+    updateData: Partial<CreateLegalRepDto>,
+  ): Promise<LegalRep>;
+}
