--- src/legalReps/strategies/create-legal-rep/create-legal-rep-strategy.interface.ts
@@ -1,6 +1,0 @@
-import { CreateLegalRepDto } from '@/legalReps/dtos/CreateLegalRepDto';
-import { LegalRep } from '@/legalReps/models/legalReps.model';
-
-export interface ICreateLegalRepsStrategy {
-  create(legalRepData: CreateLegalRepDto): Promise<LegalRep>;
-}
