--- src/legalReps/strategies/update-legal-rep/update-legal-rep-strategy.interface.ts
@@ -1,9 +1,0 @@
-import { CreateLegalRepDto } from '@/legalReps/dtos/CreateLegalRepDto';
-import { LegalRep } from '@/legalReps/models/legalReps.model';
-
-export interface IUpdateLegalRepsStrategy {
-  update(
-    legalRepId: string,
-    updateData: Partial<CreateLegalRepDto>,
-  ): Promise<LegalRep>;
-}
